week-3/references/muriel-cooper-book-designs/pyproject.toml
[project]
name = "muriel-cooper-book-designs"
version = "0.1.0"
description = "A database of Muriel Cooper's book cover designs with AI-powered scraping and a Swiss-style frontend"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.10.0",
    "aiofiles>=24.1.0",
    "pillow>=11.0.0",
]

[project.scripts]
scrape = "src.agent:main"
serve = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
